Authorizing using Gates

+++ app/Http/Controllers/QuestionController.php
@@ -72,6 +72,10 @@ class QuestionController extends Controller
     public function edit(Question $question)
     {
         //
+        if(\Gate::denies('update-question',$question)){
+
+            abort(404,'Access Denied');
+        }
         return view('questions.edit',compact('question'));
 
     }
@@ -88,6 +92,7 @@ class QuestionController extends Controller
         //
         //dd($request->all());
 
+
         $question->update($request->only('title','body'));
         return redirect()->route('questions.index')->with('success','Your Question has been Updated');
 
@@ -103,7 +108,11 @@ class QuestionController extends Controller
     {
         //
         //dd($question);
+        if(\Gate::denies('delete-question',$question)){
+            abort(404,'Access Denied');
+        }
         $question->delete();
         return redirect()->route('questions.index')->with('success','Your Question has been Deleted');
+
     }
 }
